指标点体系代码完善，后台查询效率提升

+++ iacaa2_0/iacaa2_0-admin/src/main/java/com/pzhu/iacaa2_0/entity/CourseTarget.java
@@ -28,13 +28,14 @@ public class CourseTarget extends BaseEntity {
      * 关联课程
      */
     @TableField("course_id")
-    private Integer courseId;
+    private Long courseId;
+
 
     /**
      * 关联指标点
      */
     @TableField("target_id")
-    private Integer targetId;
+    private Long targetId;
 
     /**
      * 关联比例
